requirements.txt
openai
tiktoken
requests
beautifulsoup4
python-dotenv
langchain
faiss-cpu
docx2txt
pypdf2
pypdf
python-pptx
python-multipart
tenacity